+++ packages/renderer/src/chrome/app-shell.ts
@@ -482,7 +482,7 @@ export class AppShell extends LitElement {
             target="_blank"
             rel="noopener"
             title="View the changelog on GitHub"
-            ><span class="version">v0.0.291</span></a
+            ><span class="version">v0.0.292</span></a
           ></strong
         >
         ${this.headerButtons